% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{constants}
\alias{constants}
\alias{.MariaDBPkgName}
\alias{.MariaDBPkgVersion}
\alias{.MariaDBPkgRCS}
\alias{.MariaDBSQLKeywords}
\alias{CLIENT_LONG_PASSWORD}
\alias{CLIENT_FOUND_ROWS}
\alias{CLIENT_LONG_FLAG}
\alias{CLIENT_CONNECT_WITH_DB}
\alias{CLIENT_NO_SCHEMA}
\alias{CLIENT_COMPRESS}
\alias{CLIENT_ODBC}
\alias{CLIENT_LOCAL_FILES}
\alias{CLIENT_IGNORE_SPACE}
\alias{CLIENT_PROTOCOL_41}
\alias{CLIENT_INTERACTIVE}
\alias{CLIENT_SSL}
\alias{CLIENT_IGNORE_SIGPIPE}
\alias{CLIENT_TRANSACTIONS}
\alias{CLIENT_RESERVED}
\alias{CLIENT_SECURE_CONNECTION}
\alias{CLIENT_MULTI_STATEMENTS}
\alias{CLIENT_MULTI_RESULTS}
\title{Constants}
\description{
Constants
}
\section{Constants}{
 \code{.MariaDBPkgName} (currently \code{"RMariaDB"}),
\code{.MariaDBPkgVersion} (the R package version), \code{.MariaDBPkgRCS} (the
RCS revision), \code{.MariaDBSQLKeywords} (a lot!)
}

