% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{methods}
\name{fetch,MariaDBResult,numeric-method}
\alias{fetch,MariaDBResult,numeric-method}
\title{Fetch rows (deprecated)}
\usage{
\S4method{fetch}{MariaDBResult,numeric}(res, n)
}
\description{
Please use \code{dbFetch} instead
}
\keyword{internal}
