% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MariaDBDriver.R
\name{mariadbClientLibraryVersions}
\alias{mariadbClientLibraryVersions}
\title{MariaDB Check for Compiled Versus Loaded Client Library Versions}
\usage{
mariadbClientLibraryVersions()
}
\value{
A named integer vector of length two,
the first element representing the compiled library version and the second element representing the loaded client library version.
}
\description{
This function prints out the compiled and loaded client library versions.
}
\examples{
mariadbClientLibraryVersions()
}
