% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{mariadbHasDefault}
\alias{mariadbHasDefault}
\alias{mariadbDefault}
\title{Check if default database is available.}
\usage{
mariadbHasDefault()

mariadbDefault()
}
\description{
RMariaDB examples and tests connect to a database defined by the \code{rs-dbi} group in \verb{~/.my.cnf}.
This function checks if that database is available, and if not, displays an informative message.
\code{mariadbDefault()} works similarly but throws a testthat skip condition on failure, making it suitable for use in tests.
}
\examples{
if (mariadbHasDefault()) {
  db <- dbConnect(RMariaDB::MariaDB(), dbname = "test")
  dbListTables(db)
  dbDisconnect(db)
}
}
