% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MariaDBResult.R, R/dbIsValid_MariaDBResult.R
\docType{class}
\name{MariaDBResult-class}
\alias{MariaDBResult-class}
\alias{dbIsValid_MariaDBResult}
\alias{dbIsValid,MariaDBResult-method}
\title{Class MariaDBResult}
\usage{
\S4method{dbIsValid}{MariaDBResult}(dbObj, ...)
}
\description{
MariaDB's query results class.  This classes encapsulates the result of an SQL query or statement.
}
\keyword{internal}
