% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quote.R
\docType{methods}
\name{mysql-quoting}
\alias{mysql-quoting}
\alias{dbQuoteIdentifier,MariaDBConnection,character-method}
\alias{dbQuoteString,MariaDBConnection,character-method}
\title{Quote MariaDB strings and identifiers.}
\usage{
\S4method{dbQuoteIdentifier}{MariaDBConnection,character}(conn, x, ...)

\S4method{dbQuoteString}{MariaDBConnection,character}(conn, x, ...)
}
\description{
In MariaDB, identifiers are enclosed in backticks, e.g. \code{`x`}.
}
\examples{
if (mysqlHasDefault()) {
  con <- dbConnect(RMariaDB::MariaDB())
  dbQuoteIdentifier(con, c("a b", "a`b"))
  dbQuoteString(con, c("a b", "a'b"))
  dbDisconnect(con)
}
}
\keyword{internal}
