% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMariaDB-package.R
\docType{package}
\name{RMariaDB-package}
\alias{RMariaDB}
\alias{RMariaDB-package}
\title{RMariaDB: Database Interface and MariaDB Driver}
\description{
Implements a DBI-compliant interface to MariaDB (\url{https://mariadb.org/}) and MySQL (\url{https://www.mysql.com/}) databases.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rmariadb.r-dbi.org}
  \item \url{https://github.com/r-dbi/RMariaDB}
  \item Report bugs at \url{https://github.com/r-dbi/RMariaDB/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{kirill@cynkra.com} (\href{https://orcid.org/0000-0002-1416-3412}{ORCID})

Authors:
\itemize{
  \item Kirill Müller \email{kirill@cynkra.com} (\href{https://orcid.org/0000-0002-1416-3412}{ORCID})
  \item Jeroen Ooms (\href{https://orcid.org/0000-0002-4035-0289}{ORCID})
  \item David James
  \item Saikat DebRoy
  \item Hadley Wickham
  \item Jeffrey Horner
}

Other contributors:
\itemize{
  \item R Consortium [funder]
  \item RStudio [copyright holder]
}

}
