% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{Client-flags}
\alias{Client-flags}
\title{Client flags}
\description{
Use for the \code{client.flag} argument to \code{\link[DBI:dbConnect]{dbConnect()}}, multiple flags can be
combined with \code{+} or \code{\link[=bitwOr]{bitwOr()}}.
The flags are provided for completeness.
To enforce SSL for the DB connection, add the flag \code{CLIENT_SSL}.
}
\examples{
\dontrun{
library(DBI)
library(RMariaDB)
con1 <- dbConnect(MariaDB(), client.flag = CLIENT_COMPRESS)
con2 <- dbConnect(
  MariaDB(),
  client.flag = bitwOr(CLIENT_COMPRESS, CLIENT_SSL)
)
}

}
\seealso{
The \code{flags} argument at https://mariadb.com/kb/en/library/mysql_real_connect.
}
