% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbDataType_MariaDBConnection.R,
%   R/dbDataType_MariaDBDriver.R
\name{dbDataType}
\alias{dbDataType}
\alias{dbDataType_MariaDBConnection}
\alias{dbDataType,MariaDBConnection-method}
\alias{dbDataType_MariaDBDriver}
\alias{dbDataType,MariaDBDriver-method}
\title{Determine the SQL Data Type of an S object}
\usage{
\S4method{dbDataType}{MariaDBConnection}(dbObj, obj, ...)

\S4method{dbDataType}{MariaDBDriver}(dbObj, obj, ...)
}
\arguments{
\item{dbObj}{A \linkS4class{MariaDBDriver} or \linkS4class{MariaDBConnection} object.}

\item{obj}{R/S-Plus object whose SQL type we want to determine.}

\item{\dots}{any other parameters that individual methods may need.}
}
\description{
This method is a straight-forward implementation of the corresponding
generic function.
}
\examples{
dbDataType(RMariaDB::MariaDB(), "a")
dbDataType(RMariaDB::MariaDB(), 1:3)
dbDataType(RMariaDB::MariaDB(), 2.5)
}
