% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MariaDBConnection.R
\docType{class}
\name{MariaDBConnection-class}
\alias{MariaDBConnection-class}
\alias{format.MariaDBConnection}
\alias{show,MariaDBConnection-method}
\alias{dbIsValid,MariaDBConnection-method}
\alias{dbDisconnect,MariaDBConnection-method}
\alias{dbGetInfo,MariaDBConnection-method}
\title{Class MariaDBConnection.}
\usage{
\method{format}{MariaDBConnection}(x, ...)

\S4method{show}{MariaDBConnection}(object)

\S4method{dbIsValid}{MariaDBConnection}(dbObj, ...)

\S4method{dbDisconnect}{MariaDBConnection}(conn, ...)

\S4method{dbGetInfo}{MariaDBConnection}(dbObj, what = "", ...)
}
\description{
\code{MariaDBConnection} objects are usually created by
\code{\link[DBI:dbConnect]{DBI::dbConnect()}}
}
\keyword{internal}
