% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MariaDBConnection.R,
%   R/dbDisconnect_MariaDBConnection.R, R/dbGetInfo_MariaDBConnection.R,
%   R/dbIsValid_MariaDBConnection.R, R/show_MariaDBConnection.R
\docType{class}
\name{MariaDBConnection-class}
\alias{MariaDBConnection-class}
\alias{format.MariaDBConnection}
\alias{dbDisconnect_MariaDBConnection}
\alias{dbDisconnect,MariaDBConnection-method}
\alias{dbGetInfo_MariaDBConnection}
\alias{dbGetInfo,MariaDBConnection-method}
\alias{dbIsValid_MariaDBConnection}
\alias{dbIsValid,MariaDBConnection-method}
\alias{show_MariaDBConnection}
\alias{show,MariaDBConnection-method}
\title{Class MariaDBConnection.}
\usage{
\method{format}{MariaDBConnection}(x, ...)

\S4method{dbDisconnect}{MariaDBConnection}(conn, ...)

\S4method{dbGetInfo}{MariaDBConnection}(dbObj, what = "", ...)

\S4method{dbIsValid}{MariaDBConnection}(dbObj, ...)

\S4method{show}{MariaDBConnection}(object)
}
\description{
\code{MariaDBConnection} objects are usually created by
\code{\link[DBI:dbConnect]{DBI::dbConnect()}}
}
\keyword{internal}
