% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MariaDBDriver.R
\docType{class}
\name{MariaDBDriver-class}
\alias{MariaDBDriver-class}
\alias{dbUnloadDriver,MariaDBDriver-method}
\alias{dbIsValid,MariaDBDriver-method}
\alias{dbGetInfo,MariaDBDriver-method}
\title{Class MariaDBDriver with constructor MariaDB.}
\usage{
\S4method{dbUnloadDriver}{MariaDBDriver}(drv, ...)

\S4method{dbIsValid}{MariaDBDriver}(dbObj, ...)

\S4method{dbGetInfo}{MariaDBDriver}(dbObj, ...)
}
\description{
An MariaDB driver implementing the R database (DBI) API.
This class should always be initialized with the \code{\link[=MariaDB]{MariaDB()}} function.
It returns a singleton that allows you to connect to MariaDB.
}
\keyword{internal}
