% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbQuoteLiteral_MySQLConnection.R,
%   R/quote-mysql.R
\name{mysql-quoting}
\alias{mysql-quoting}
\alias{dbQuoteLiteral_MySQLConnection}
\alias{dbQuoteLiteral,MySQLConnection-method}
\title{Quote MySQL strings and identifiers.}
\usage{
\S4method{dbQuoteLiteral}{MySQLConnection}(conn, x, ...)
}
\description{
In MySQL, identifiers are enclosed in backticks, e.g. \code{`x`}.
}
\keyword{internal}
