% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMariaDB.R
\docType{package}
\name{RMariaDB-package}
\alias{RMariaDB}
\alias{RMariaDB-package}
\title{RMariaDB: Database Interface and 'MariaDB' Driver}
\description{
Implements a 'DBI'-compliant interface to 'MariaDB' (<https://mariadb.org/>) and 'MySQL' (<https://www.mysql.com/>) databases.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstats-db/RMariaDB}
  \item \url{https://downloads.mariadb.org/connector-c/}
  \item Report bugs at \url{https://github.com/rstats-db/RMariaDB/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org}

Authors:
\itemize{
  \item Jeroen Ooms
  \item David James
  \item Saikat DebRoy
  \item Hadley Wickham
  \item Jeffrey Horner
}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
  \item Kungliga Tekniska Högskolan (Source code for timegm) [contributor]
}

}
